Fix GCC-15 build failures (Closes: #1097375)
The 'original' upstream on sourceforge seems to be abandoned, but 'John
Brandwood' made a fork on GitHub where the GCC 14 fix also came from.

In their repo an issue was created for the GCC 15 issue and he provided
a fix in that issue, but also committed it to their repo in commit
c326468d6c54 ("Fix configure.ac failing to define LL_FORM when using GCC v15.")

As that fork also applied other changes, including applying (modified
versions of) Debian patches, adjust the context of the patch so that it
applies to what Debian currently has.

Next to the direct fix to configure.ac which prevented the build from
starting all together, GCC 15 is also more strict and it doesn't like
labels with value 'true' or 'false'.
That issue is also fixed in the fork with commit
1d65736ca2c1 ("Stop using "true" and "false" as label names in system.c, it breaks modern C!")

Link: https://bugs.debian.org/1097375
Link: https://github.com/jbrandwood/mcpp/issues/1
Link: https://github.com/jbrandwood/mcpp/commit/c326468d6c5437e0c846eabbc2f5cf501143d66d
Link: https://github.com/jbrandwood/mcpp/commit/1d65736ca2c10328916ed6c98e6422a784da187a
Signed-off-by: Diederik de Haas <didi.debian@cknow.org>